Add comment and rating entry tables

+++ Business/Dto/OutputDto/BrandOutMultiDto.cs
@@ -6,8 +6,8 @@
         public string BrandName { get; set; }
         public string CompanyName { get; set; }
         public double RatingTotal { get; set; }
-        public int RatingAnimals { get; set; }
-        public int RatingPeople { get; set; }
-        public int RatingPlanet { get; set; }
+        public double RatingAnimals { get; set; }
+        public double RatingPeople { get; set; }
+        public double RatingPlanet { get; set; }
     }
 }
